Changed EN behaviour to match updated motor board
In new board, enable is used in hardware using an AND gate.
Therefore EN is needed to move the motor.
If moving manually, the code writes the enable manually. Otherwise, it is given by external board.
This is fine since there is a series resistor from external EN signal: there is never a short circuit between an external EN signal and the digitalWrite from the motor board itself.